requirements.txt
pydantic~=2.9.2
fastapi~=0.115.4
tortoise-orm~=0.21.7
fastapi-jwt~=0.3.0
authlib~=1.3.0
aerich~=0.7.2
python-multipart~=0.0.17